app/requirements.txt
flask
flask_restful
flask_cors
flask-sqlalchemy
flask-migrate
celery
python-dotenv
pandas
glm==0.4.3
h5py
Keras==2.6.0
numpy
open3d==0.14.1
opencv_python==4.5.5.62
Pillow==9.0.1
PyOpenGL==3.1.6
PySide2==5.15.2.1
requests==2.27.1
scikit_learn==1.0.2
scipy==1.7.3
scikit-image
tensorflow-gpu==2.6.0
tritonclient[all]
pixellib
protobuf==3.20.0